backend/requirements.txt
django==4.2.9
environs[django]==11
pillow==11
django-debug-toolbar==4.4.6
django-phonenumber-field[phonenumberslite]==8.0.0
djangorestframework==3.15.2
requests==2.32.3
geopy==2.4.1
psycopg2-binary==2.9.10
requests==2.32.*
gunicorn==23.0.0
rollbar==1.2.0
